backend/api/requirements.txt
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
bcrypt==4.1.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==3.10.1
django-redis==5.4.0
django_pg8000==0.0.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.6
joblib==1.3.2
luas==0.0.1a0
luas.py==0.3.3
numpy==1.26.2
packaging==23.2
pandas==1.5.3
prophet==1.1.5
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
scramp==1.4.4
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
typing_extensions==4.8.0
urllib3==2.2.0
